requirements.txt
flask
opencv-python
numpy
requests
pillow
scikit-learn
python-dotenv